requirements.txt
click==8.1.3
greenlet==2.0.1
ujson==5.6.0
fastapi==0.88.0
gunicorn==20.1.0
uvicorn==0.20.0
pythondi==1.2.4 
PyJWT==2.6.0
prisma==0.7.1
fastapi_keycloak==1.0.7
prometheus-fastapi-instrumentator==5.9.1